Fix/improve-error-messages

+++ src/main/java/seedu/address/logic/commands/DeletePersonFromTableCommand.java
@@ -9,6 +9,8 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.table.Table;
 
 /**
  * Deletes a guest from the given table.
@@ -28,8 +30,6 @@ public class DeletePersonFromTableCommand extends Command {
         + PREFIX_TABLE_ID + "2";
 
     public static final String MESSAGE_REMOVED_GUEST_FROM_TABLE_SUCCESS = "Deleted Person: %s from Table: %d";
-    public static final String MESSAGE_NOT_EDITED = "At least one field to edit must be provided.";
-    public static final String MESSAGE_DUPLICATE_PERSON = "This guest already exists in the address book.";
 
     private final Name guestName;
     private final int oldTableId;
@@ -50,13 +50,45 @@ public class DeletePersonFromTableCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
 
-        Person guestToRemove = model.findPersonByName(guestName);
+        // 1. Check if wedding is set
+        if (!model.hasCurrentWedding()) {
+            throw new CommandException("No current wedding set. Please use 'setWedding' first.");
+        }
 
+        // 2. Try to find the person by name, catch if not found
+        Person guestToRemove;
+        try {
+            guestToRemove = model.findPersonByName(guestName);
+        } catch (seedu.address.model.person.exceptions.PersonNotFoundException e) {
+            throw new CommandException(String.format("Person '%s' not found in the guest list.", guestName.fullName));
+        }
+
+        // 3. Check if the table exists
+        if (!model.hasTable(oldTableId)) {
+            throw new CommandException(String.format("Table with ID %d does not exist.", oldTableId));
+        }
+
+        Table table = model.getCurrentWedding().findTableById(oldTableId);
+        if (table == null) {
+            throw new CommandException(String.format("Table with ID %d could not be found.", oldTableId));
+        }
+
+        // 4. Check if the person is assigned to the table
+        try {
+            table.findPerson(guestToRemove); // throws if not assigned
+        } catch (PersonNotFoundException e) {
+            throw new CommandException(String.format("Person '%s' is not assigned to Table %d.",
+                    guestName.fullName, oldTableId));
+        }
+
+        // 5. Safe to delete
         model.deletePersonFromTableById(guestToRemove, oldTableId);
 
         return new CommandResult(String.format(MESSAGE_REMOVED_GUEST_FROM_TABLE_SUCCESS,
-            guestToRemove.getName().fullName, oldTableId));
+                guestToRemove.getName().fullName, oldTableId));
     }
+
+
 
     @Override
     public boolean equals(Object other) {
